+++ integration_test/va.test.js
@@ -2,25 +2,33 @@ const x = require('./xendit.test');
 
 const VirtualAcc = x.VirtualAcc;
 const va = new VirtualAcc({});
-
+function sleepFor(sleepDuration) {
+  var now = new Date().getTime();
+  while (new Date().getTime() < now + sleepDuration) {
+    /* Do nothing */
+  }
+}
 module.exports = function() {
   return va
     .getVABanks()
     .then(banks => {
       return va.createFixedVA({
-        externalID: '123',
+        externalID: 'VA-xendit-node-js998877',
         bankCode: banks[0].code,
         name: 'Stanley Nguyen',
         isClosed: true,
         expectedAmt: 10000,
       });
     })
-    .then(({ id }) => va.getFixedVA({ id }))
     .then(({ id }) => {
+      sleepFor(3000);
+      return va.getFixedVA({ id });
+    })
+    .then(({ id }) => {
+      sleepFor(5000);
       return va.updateFixedVA({
         id,
-        suggestedAmt: 10000,
-        expectedAmt: 10000,
+        expectedAmt: 12000,
       });
     })
     .then(() => {