+++ src/card/charge.d.ts
@@ -12,6 +12,8 @@ export function createCharge(data: {
   promotion?: object;
   installment?: object;
   forUserID?: string;
+  metadata?: object;
+  isRecurring: boolean;
 }): Promise<object>;
 
 export function captureCharge(data: {