+++ src/recurring_payment/manage_payments.d.ts
@@ -1,17 +1,18 @@
-enum Interval {
+export enum Interval {
   Day = 'DAY',
   Week = 'WEEK',
   Month = 'MONTH',
 }
-enum Action {
+
+export enum Action {
   Stop = 'STOP',
   Ignore = 'IGNORE',
 }
 
 export function createPayment(data: {
   externalID: string;
-  payerEmail: string;
-  description: string;
+  payerEmail?: string;
+  description?: string;
   amount: number;
   interval: Interval;
   intervalCount: number;
@@ -27,6 +28,11 @@ export function createPayment(data: {
   chargeImmediately?: boolean;
   currency?: string;
   rescheduleAt?: Date;
+  customer?: object;
+  customerNotificationPreference?: object;
+  reminderTimeUnit?: string;
+  reminderTime?: number;
+  paymentMethodId?: string;
 }): Promise<object>;
 
 export function getPayment(data: { id: string }): Promise<object>;
@@ -41,4 +47,8 @@ export function editPayment(data: {
   invoiceDuration?: number;
   missedPaymentAction?: Action;
   rescheduleAt?: Date;
+  customerId?: string;
+  reminderTimeUnit?: string;
+  reminderTime?: number;
+  paymentMethodId?: string;
 }): Promise<object>;